:recycle: to separate components of BestCombinationListPage
